fix(tests): split the orphan heartbeat test file to satisfy PS-204 mirroring
CI's scitex-dev audit (tests/develop/test_audit.py::test_audit_all_clean)
failed with ONE error — no test regression, 9628 tests passed:

  [E] [PS-204 §2 orphan-test-file]
      tests/.../_lifecycle/test__heartbeat_unknown_is_not_dead.py:
      no matching src file (orphan test)

PS-204 requires each test file's stem to resolve to a real source module
in the mirrored src dir. `test__heartbeat_unknown_is_not_dead.py` expects
`_lifecycle/_heartbeat_unknown_is_not_dead.py`, which does not exist — it
was a cross-cutting file covering BOTH heartbeat loops, so it mirrored
neither.

The rule is a PREFIX match, not an exact one — verified against the repo
rather than assumed: 107 existing test files have no exact mirror (so an
exact-match rule would already redden develop, which is green), while
exactly ONE had no prefix match — this file. That is a 1:1 match with the
audit reporting exactly 1 error, on the same path.

Split it along the loop it actually exercises, so each half mirrors a real
module by prefix:

  test__tui_heartbeat_loop_unknown_is_not_dead.py -> _tui_heartbeat_loop.py
  test__sdk_heartbeat_loop_unknown_is_not_dead.py -> _sdk_heartbeat_loop.py

The TUI half keeps the real-state.db reachability tests (the loop they
drive is the TUI one). No test was weakened, dropped, or skipped — the
same 41 assertions still run, and the UNKNOWN-is-not-DEAD invariant is
unchanged. Both files are well under the 512-line cap (240 / 177).

Re-verified: 0 orphan test files repo-wide; 41 passed.
